backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
httpx==0.28.1
supabase==2.28.3
python-dotenv==1.2.2
apscheduler==3.11.2
pydantic==2.12.5
